requirements.txt
pandas
sqlalchemy
google-generativeai
python-dotenv
psycopg2-binary
streamlit
